Pedro's changes Dec 18. OK, so this is what I've done so far:
As metioned earlier - separated the HUD glass part from the cockpit_INN object. Stopped using the cockpit_OUT object and rearranged the order of all misc. objects (logical, for transparency en/or lighting). Od GPS and FMS replaced by the xp11 versions G430 and FMC, as 2D PlaneMaker instruments placen on the 3D panel. Adjustments to the landing gear legs' length to have the (3D) tires now touching the ground.

If you look in the cockpit/-PANELS- folder you'll see several Panel.png's, these are different versions of the HUD glass ranging from fully transparent (white) to light, medium and dark greenish (different transparency); 'medium' is installed -to install/test a different one just rename the files as needed before (re)strating x-plane, Pnel.png is the one in use.

Installed interior lighting to the cockpit (PlaneMaker - Standard/Viewpoint/Interior Lights) for those wanting to fly other that in daylight and nice weather... And adjusted the intensity of the landing- and taxi lights a bit.

Setup store positions for all pylons; that is all except for the RH outer one where twin-rail Sidewinder rack is and the 2 pylons with Mavericks still have the the extra rail objects in place. That said, the obj. and png. for the Maverick launcher/rail have been renamed to what they probably were meant to be: LAU-117. see pic as a sample of the locations (most of them); starting at the top with the gun, then the center pylon and then the inner pylons first Right then Left and the next pylons moving on to the outer ones down the list. Oh yes, adjusted gun (weight, rate of fier etc.) and new sound (free from internet).

Some of the weapons were revised (see pic, purple for modified, red for the 'blank' placeholder wpn). Adjustmens are mostly to the attach point of the weapon and sometimes also to weight. NEW the Sniper ATP is now also setup as weapon (built a wpn for it and moved the obj and png to the weapons forlder for use), that way  you can also mout or remove it at will -like the ALQ-131 ECM pod as a weapons it doesn't a thing, it's just a zero range laser and can't be dropped either. Built and installed a weapon for Chaff (wingtips) and Flares (main gear pods).

The way the store locations and weapons are now setup means you can now re-arm with other weapons (providing they were built the right way, xp11 weapons and not earlier xp10 or 9).

That said, I may or may not go on make a few more adjustments... First thing up is a quick mod removing the LAU-117 objects and adjusting the pylon's attach point; it may then not look like the real think but that way we can also re-arm these locations with other weapons.  Later maybe convert both the LAU-117 and LAU-114A launchers to weapons for easy mount or remove with different load outs.

Have fun!

Cheers,  Pedro

PS the file itself was too large (55.9mb) to upload here, so you can downlod it from here (starts direct download of zip file from my site)